fix(dashboard): validate project id as integer in schema
Add `.int()` to projectBoardInputSchema to reject NaN and
floating-point values from route params.

+++ apps/agent-please/server/orpc/schemas.ts
@@ -135,7 +135,7 @@ export const projectsResponseSchema = z.object({
 })
 
 export const projectBoardInputSchema = z.object({
-  id: z.number().nonnegative(),
+  id: z.number().int().nonnegative(),
 })
 
 export const boardIssueSchema = z.object({
